When search does not provide any results indicates it. Refactored all components in the project to be in their own component and got all imports to be named imports
