+++ .gitignore
@@ -24,3 +24,6 @@ dist-ssr
 *.sw?
 
 *.env
+
+# Local Netlify folder
+.netlify
