+++ src/App.tsx
@@ -1,12 +1,12 @@
 import "./scss/style.scss";
-import Header from "./components/Header/Header.tsx";
-import Footer from "./components/Footer/Footer.tsx";
+import { Header } from "./components/Header/Header.tsx";
+import { Footer } from "./components/Footer/Footer.tsx";
 import { Outlet, useLocation } from "react-router-dom";
 import { useEffect } from "react";
 import { SearchContextProvider } from "./contexts/SearchContextProvider.tsx";
 import { ActiveFilterContextProvider } from "./contexts/ActiveFilterContextProvider.tsx";
 
-function App() {
+export const App = () => {
   const location = useLocation();
 
   useEffect(() => {
@@ -26,6 +26,4 @@ function App() {
       </ActiveFilterContextProvider>
     </>
   );
-}
-
-export default App;
+};
