+++ src/components/ItemsToDisplay/ItemsToDisplay.tsx
@@ -1,47 +1,16 @@
 import "./ItemsToDisplay.scss";
 import { Link } from "react-router-dom";
 import { ItemDetailsModel } from "../../models";
-import { useState, useEffect, useMemo } from "react";
-import { useContext } from "react";
-import { SearchContext } from "../../contexts/SearchContextProvider.tsx";
-import { useShopItemsList } from "../../services/useShopItemsList.ts";
-import { ActiveFilterContext } from "../../contexts/ActiveFilterContextProvider.tsx";
 
-const ItemsToDisplay = () => {
-  const { shopItemsList } = useShopItemsList();
-  const [search, setSearch] = useState<string>("");
-  const { search: searchValue } = useContext(SearchContext);
-  const { activeFilter } = useContext(ActiveFilterContext);
+interface ItemsToDisplayProps {
+  itemsList: ItemDetailsModel[];
+}
 
-  // Get search value from search bar on header
-  useEffect(() => {
-    setSearch(searchValue);
-  }, [searchValue]);
-
-  const displayedItems = useMemo(() => {
-    if (!search && !activeFilter) {
-      return shopItemsList;
-    }
-
-    return shopItemsList.filter((item: ItemDetailsModel) => {
-      // If there is an active filter and the item does not match with this active filter, return false
-      if (activeFilter && activeFilter !== item.origin) {
-        return false;
-      }
-
-      if (!search) {
-        return true;
-      }
-
-      // Return true for items that include in its name the search value
-      return item.name.toLowerCase().includes(search.toLowerCase());
-    });
-  }, [search, activeFilter, shopItemsList]);
-
+export const ItemsToDisplay = ({ itemsList }: ItemsToDisplayProps) => {
   return (
     <>
       <div className={"itemDisplay"}>
-        {displayedItems.map((item: ItemDetailsModel) => (
+        {itemsList.map((item: ItemDetailsModel) => (
           <Link
             className={"item-link"}
             to={`/detalles-producto/?ref=${item.key}`}
@@ -60,5 +29,3 @@ const ItemsToDisplay = () => {
     </>
   );
 };
-
-export default ItemsToDisplay;
