+++ src/components/ProductDetailsComponents/Review.tsx
@@ -1,0 +1,16 @@
+export const Review = () => {
+  return (
+    <div>
+      <div>
+        <img src={"../public/user.png"} alt={""} />
+        <span>Pepito</span>
+      </div>
+      <p>
+        Review: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
+        eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
+        minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip
+        ex ea commodo consequat.
+      </p>
+    </div>
+  );
+};
