+++ src/scss/utilities/_positioning.scss
@@ -41,6 +41,10 @@
   margin-left: auto;
 }
 
+.margin-left-60 {
+  margin-left: 60px;
+}
+
 .margin-bottom-0 {
   margin-bottom: 0;
 }
