+++ src/scss/wrappers/MainWrapper.tsx
@@ -1,18 +1,18 @@
-import Header from '../../components/Header/Header.tsx'
-import Footer from '../../components/Footer/Footer.tsx'
+import { Header } from "../../components/Header/Header.tsx";
+import { Footer } from "../../components/Footer/Footer.tsx";
 
 type Props = {
-    children?: React.ReactNode
+  children?: React.ReactNode;
 };
 
-export const MainWrapper = ({children}: Props): React.ReactNode => {
-    return (
-        <>
-            <Header />
-            {children}
-            <Footer />
-        </>
-    );
+export const MainWrapper = ({ children }: Props): React.ReactNode => {
+  return (
+    <>
+      <Header />
+      {children}
+      <Footer />
+    </>
+  );
 };
 
-export default MainWrapper
+export default MainWrapper;
